minor restructure of functionality, tested with sample data, very spicy so far
